Added code changes for terminating application on windows/linux without commenting code
git-svn-id: https://openmalaria.googlecode.com/svn@1499 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ application_deployment/start_gui.py
@@ -41,9 +41,11 @@ class OMFrontend:
     destroy:
     Kills the actual window'''
     def destroy(self, widget, data=None):
-        #gtk.main_quit()
-        #sys.exit()
-	    os.kill(os.getpid(),signal.SIGTERM)
+        if sys.platform == 'win32':
+            gtk.main_quit()
+            sys.exit()
+        else:
+            os.kill(os.getpid(),signal.SIGTERM)
         
     def __init__(self):
         self.window = gtk.Window(gtk.WINDOW_TOPLEVEL)
